Refactor lead/poc props into base data model
The ipac_lead and spherex_poc attributes are Python properties rather
than JSON serialized data (they are based on finding authors with roles
corresponding to "IPAC Lead" or "SPHEREx Lead" respectively, which in
turn are set when parsing the \ipaclead and \spherexlead LaTeX commands
in the base parser). This means it isn't necessary after all to version
those attributes since they're internal to the Lander plugin (although
the document portal copies that code; so we'll need to update that as
well.

To make this more maintainable, this commit refactors these properties
down into the base SpherexMetadata class. This works out well because
the documents share their HTML template, so the meaning of IPAC Lead,
SPHEREx POC, and other authors is consistent.